feat: add useGifs custom hook
